U,D-예외처리 추가

+++ backend/src/main/java/hongsam/demo/board/service/BoardService.java
@@ -4,13 +4,12 @@ import hongsam.demo.board.domain.Board;
 import hongsam.demo.board.domain.BoardMemberResponse;
 import hongsam.demo.board.domain.BoardResponse;
 import hongsam.demo.board.repository.BoardRepository;
-import hongsam.demo.board.repository.BoardUpdateDto;
+import hongsam.demo.board.domain.BoardUpdateDto;
 import hongsam.demo.member.domain.MemberDto;
 import hongsam.demo.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 import java.util.Optional;
 
@@ -46,12 +45,49 @@ public class BoardService {
     }
 
 
-    public void updateBoard(Long boardId, BoardUpdateDto boardUpdateDto) {
+    public BoardResponse updateBoard(Long boardId, BoardUpdateDto boardUpdateDto) {
+
+        // 회원 존재 확인
+        Optional<MemberDto> findMember = memberRepository.findById(boardUpdateDto.getMemberId());
+        if (!findMember.isPresent()) {
+            return new BoardResponse(400, "존재하지 않는 회원입니다.");
+        }
+        Optional<Board> board = boardRepository.getBoardById(boardId);
+        // 보드 존재 확인
+        if (!board.isPresent()) {
+            return new BoardResponse(400, "존재하지 않는 게시물입니다.");
+        }
+        // 넘어온 memberId != 보드의 memberId
+        if (boardUpdateDto.getMemberId() != board.get().getMemberId()) {
+            return new BoardResponse(400, "게시물의 소유자가 아닙니다.");
+        }
+
         boardRepository.updateBoard(boardId,boardUpdateDto);
+
+        return new BoardResponse(200, "게시물이 성공적으로 수정되었습니다.");
     }
 
-    public void deleteBoard(Long boardId) {
+    public BoardResponse deleteBoard(Long boardId, Long memberId) {
+
+        // 회원 존재 확인
+        Optional<MemberDto> findMember = memberRepository.findById(memberId);
+        if (!findMember.isPresent()) {
+            return new BoardResponse(400, "존재하지 않는 회원입니다.");
+        }
+
+        Optional<Board> board = boardRepository.getBoardById(boardId);
+        // 보드 존재 확인
+        if (!board.isPresent()) {
+            return new BoardResponse(400, "존재하지 않는 게시물입니다.");
+        }
+        // 넘어온 memberId != 보드의 memberId
+        if (memberId != board.get().getMemberId()) {
+            return new BoardResponse(400, "게시물의 소유자가 아닙니다.");
+        }
+
         boardRepository.deleteBoard(boardId);
+
+        return new BoardResponse(200, "게시물이 성공적으로 삭제되었습니다.");
     }
 
 }
